+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v1.3.0
  - Remove unused `pristineValue` property in `Field`.
  - Add TypeScript type definitions.
  - Add `moveUp` and `moveDown` to `ListForm`.